Added cast.

+++ distance.c
@@ -4,5 +4,7 @@ int distance(
 	const struct Coordinate *first,
 	const struct Coordinate *second)
 {
-	return (int)fmax(abs(first->row - second->row), abs(first->col - second->col));
+	return (int)fmax(
+		(double)abs(first->row - second->row),
+		(double)abs(first->col - second->col));
 }
